adding GetMPICommFromEpetraComm
