+++ applications/TrilinosApplication/custom_utilities/trilinos_solver_utilities.h
@@ -7,7 +7,7 @@
 //  License:		 BSD License
 //					 Kratos default license: kratos/license.txt
 //
-//  Main authors:    Philipp Bucher
+//  Main authors:    Philipp Bucher (https://github.com/philbucher)
 //
 
 #if !defined (KRATOS_TRILINOS_SOLVER_UTILITIES_H_INCLUDED)
@@ -15,6 +15,8 @@
 
 // External includes
 #include "Teuchos_ParameterList.hpp"
+#include <mpi.h>
+#include "Epetra_Comm.h"
 
 // Project includes
 #include "includes/define.h"
@@ -23,15 +25,9 @@
 namespace Kratos {
 namespace TrilinosSolverUtilities {
 
-void SetTeuchosParameters(const Parameters rSettings, Teuchos::ParameterList& rParameterlist)
-{
-    for (auto it = rSettings.begin(); it != rSettings.end(); ++it) {
-        if      (it->IsString()) rParameterlist.set(it.name(), it->GetString());
-        else if (it->IsInt())    rParameterlist.set(it.name(), it->GetInt());
-        else if (it->IsBool())   rParameterlist.set(it.name(), it->GetBool());
-        else if (it->IsDouble()) rParameterlist.set(it.name(), it->GetDouble());
-    }
-}
+void SetTeuchosParameters(const Parameters rSettings, Teuchos::ParameterList& rParameterlist);
+
+MPI_Comm GetMPICommFromEpetraComm(const Epetra_Comm&);
 
 }  // namespace TrilinosSolverUtilities.
 }  // namespace Kratos.
